refactor(profile): handle and persistence refactor
* fix(profile): switch primary key from account_id to id and rename account_id to owner_id

* fix(profile): rename profile username to handle

* refactor(profile): decouple profile identity from account identity

BREAKING CHANGE: Profile now uses its own ProfileId instead of AccountId as Primary Key.
This enables 1:N relationship between Accounts and Profiles.

Changes:
- Domain: Introduced ProfileId ValueObject and updated Profile aggregate.
- Domain: Renamed 'username' to 'handle' to align with social domain language.
- Persistence: Migrated PostgreSQL schema to use (id, region_code) as Composite PK.
- Persistence: Unified Identity, Stats, and Cache via UnifiedProfileRepository.
- Kernel: Added Slug ValueObject for shared validation logic (Handle/Username).
- Infrastructure: Updated Kafka consumer to handle account.created events with the new identity mapping.
- API: Updated gRPC mappers and handlers to reflect handle/id changes.feat(profile): add post_count to profile stats with ScyllaDB integration

- Update ProfileStats domain value object to include post_count.
- Add post_count field to profile.v1.ProfileStats protobuf message.
- Update ScyllaProfileRepository to fetch and save post_count using counters.
- Update gRPC mappers to handle the new field.
- Refactor integration tests structure for better stability with Redis.

Architecture decision: post_count is stored in ScyllaDB alongside other
social counters to ensure write scalability and consistent data locality
within the CompositeProfileRepository.

---------

+++ crates/account/src/domain/builders/account_metadata_builder.rs
@@ -12,7 +12,7 @@ pub struct AccountMetadataBuilder {
     role: AccountRole,
     trust_score: i32,
     estimated_ip: Option<String>,
-    version: i32,
+    version: u64,
 }
 
 impl AccountMetadataBuilder {
@@ -41,7 +41,7 @@ impl AccountMetadataBuilder {
         moderation_notes: Option<String>,
         estimated_ip: Option<String>,
         updated_at: DateTime<Utc>,
-        version: i32,
+        version: u64,
     ) -> AccountMetadata {
         AccountMetadata::restore(
             account_id,
